hentFagsakerForPerson gjrøes kun når den trengs, og ikke hver gang man henter en bruker
* Tatt ut hentFagsakerForPerson i egen fil
* Modalen henter kun fagsaker fra saksoversikten idag, og har beholdt det
