+++ src/frontend/context/fagsak/useHentFagsakerForPerson.ts
@@ -1,0 +1,23 @@
+import { useHttp } from '@navikt/familie-http';
+import type { Ressurs } from '@navikt/familie-typer';
+
+import type { IMinimalFagsak } from '../../typer/fagsak';
+
+export const useHentFagsakerForPerson = () => {
+    const { request } = useHttp();
+    const hentFagsakerForPerson = async (personId: string) => {
+        return request<{ personIdent: string }, IMinimalFagsak[]>({
+            method: 'POST',
+            url: `/familie-ba-sak/api/fagsaker/hent-fagsaker-paa-person`,
+            data: {
+                personIdent: personId,
+            },
+        }).then((fagsaker: Ressurs<IMinimalFagsak[]>) => {
+            return fagsaker;
+        });
+    };
+
+    return {
+        hentFagsakerForPerson,
+    };
+};
