+++ src/frontend/komponenter/Fagsak/Personlinje/Behandlingsmeny/OpprettFagsak/OpprettFagsak.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import { Dropdown } from '@navikt/ds-react-internal';
 
-import { useFagsakContext } from '../../../../../context/fagsak/FagsakContext';
 import type { IPersonInfo } from '../../../../../typer/person';
 import OpprettFagsakModal from '../../../../Felleskomponenter/HeaderMedSøk/OpprettFagsakModal';
 
@@ -12,7 +11,6 @@ interface IProps {
 
 const OpprettFagsak: React.FC<IProps> = ({ personInfo }) => {
     const [visModal, settVisModal] = React.useState(false);
-    const { fagsakerPåBruker } = useFagsakContext();
 
     return (
         <>
@@ -22,7 +20,6 @@ const OpprettFagsak: React.FC<IProps> = ({ personInfo }) => {
             {visModal && (
                 <OpprettFagsakModal
                     personInfo={personInfo}
-                    fagsakerPåBruker={fagsakerPåBruker}
                     lukkModal={() => {
                         settVisModal(false);
                     }}
